Changed requestTimeout's type to java.time.Duration (#1864)

+++ client/src/main/java/org/asynchttpclient/netty/timeout/RequestTimeoutTimerTask.java
@@ -29,7 +29,7 @@ public class RequestTimeoutTimerTask extends TimeoutTimerTask {
     RequestTimeoutTimerTask(NettyResponseFuture<?> nettyResponseFuture,
                             NettyRequestSender requestSender,
                             TimeoutsHolder timeoutsHolder,
-                            int requestTimeout) {
+                            long requestTimeout) {
         super(nettyResponseFuture, requestSender, timeoutsHolder);
         this.requestTimeout = requestTimeout;
     }
